fix: install all public headers and pin kdtree to a release
The FILE_SET HEADERS list carried 6 of the 13 headers under include/, so
`cmake --install` produced a tree that could not be compiled against:
even the documented entry point failed, since mrmr/mrmr.hpp includes
mrmr/dataset.hpp which includes the uninstalled mrmr/mi_policy.hpp. No
CI job could see this because every one of them consumes the headers
through the library's BUILD_INTERFACE, straight out of the source tree.

Add an `install` job that installs to a staging prefix, asserts the
installed header set matches include/, and builds test/install-test
against that prefix alone via find_package(mrmr).

The optional kdtree dependency was fetched from a moving `develop`
branch over `git@` SSH — non-reproducible, and unfetchable by CI or by
any consumer without a key for the Gitea host. Pin it to release v3.0.0
via MRMR_KDTREE_VERSION, sourced from the public mirror.

MRMR_CONTINUOUS was likewise never built or tested in CI despite being
documented, and had accumulated two -Wconversion errors that made any
such build fail under -Werror. Fix those and add a `continuous` preset
and CI job running the existing KSG suite (62 tests vs 43) under GCC and
Clang.

Note that an installation configured with MRMR_CONTINUOUS=ON exports a
dependency on kdtree::kdtree that FetchContent never installs, so it is
not consumable via find_package. That is left as-is and documented; the
`install` job covers the default configuration.

+++ test/bench_continuous.cpp
@@ -142,7 +142,7 @@ TEST_CASE("bench: mixed_dataset MI dispatch", "[!benchmark][continuous]") {
     std::mt19937 gen(42);
     std::vector<double> data(10000 * 3);
     for (std::size_t i = 0; i < 10000; ++i) {
-      data[i * 3 + 0] = (i % 4);
+      data[i * 3 + 0] = static_cast<double>(i % 4);
       data[i * 3 + 1] = std::normal_distribution<double>(data[i * 3 + 0], 1.0)(gen);
       data[i * 3 + 2] = std::normal_distribution<double>(0.0, 1.0)(gen);
     }
